W2_Agentic_Workflow/app/requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
openai>=1.50.0
streamlit>=1.40.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
tenacity>=8.0.0
folium>=0.17.0
streamlit-folium>=0.22.0
plotly>=5.22.0
weasyprint>=62.0
qrcode[pil]>=7.0
jinja2>=3.1.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
pytest>=8.0.0
tiktoken>=0.7.0
python-dateutil>=2.8.0
